Extract some small components we can reuse in the future
