+++ public/index.html
@@ -3,7 +3,6 @@
 <head>
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link rel="stylesheet" type="text/css" href="%PUBLIC_URL%/css/index.css">
   <title>Buzz</title>
 </head>
 <body>
