+++ src/App.js
@@ -4,6 +4,7 @@ import LogInScreen from "./session/LogInScreen";
 import { HashRouter, Route } from "react-router-dom";
 import { ALBUM, INDEX, LOG_IN } from "./routes";
 import AlbumScreen from "./album/AlbumScreen";
+import "./App.css";
 
 export default class App extends Component {
   render() {
