+++ src/splashScreen/test/splashScreen.test.js
